Move the IndexScheduler to the root of the index-scheduler crate
